# cmd/vmihub/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is vmihub server.",
        "title": "vmihub project",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/image/:username/:name/startChunkUpload": {
            "post": {
                "description": "UploadImageChunk upload image chunk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "upload image chunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "镜像名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "强制上传（覆盖）",
                        "name": "force",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "chunk大小",
                        "name": "chunkSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "chunk数量",
                        "name": "nChunks",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/image/chunk/merge": {
            "post": {
                "description": "MergeChunk merge chunk slice file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "merge chunk slice file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传uploadID",
                        "name": "uploadID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/image/chunk/{chunkIdx}/upload": {
            "post": {
                "description": "UploadImageChunk upload image chunk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "upload image chunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分片序列",
                        "name": "chunkIdx",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传uploadID",
                        "name": "uploadID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/image/{username}/{name}/chunk/{chunkIdx}/download": {
            "get": {
                "description": "DownloadImageChunk download image chunk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "镜像管理"
                ],
                "summary": "download image chunk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "仓库用户名",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "仓库名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分片序号",
                        "name": "chunkIdx",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"latest\"",
                        "description": "标签",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"50M\"",
                        "description": "分片大小",
                        "name": "chunkSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}